Refactor : 쓰레즈 멀티 레이어 추가
