+++ app.js
@@ -1,17 +1,21 @@
 const express = require("express");
-const http = require("http");
 const cors = require("cors");
 const morgan = require("morgan");
 const dotenv = require("dotenv");
-const threadsService2 = require('./heywoo/threadsService2.js')
-
 dotenv.config();
+const jwt = require("jsonwebtoken");
+const { myDataSource } = require("./src/models/dataSource");
+const routers = require("./src/routers");
+const { errorHandler } = require("./src/utils/errorHandler");
 
 const app = express();
 
 app.use(cors());
 app.use(morgan("dev"));
 app.use(express.json());
+app.use(routers);
+
+app.use(errorHandler);
 
 app.get("/ping", async (req, res) => {
   try {
@@ -23,11 +27,12 @@ app.get("/ping", async (req, res) => {
   }
 });
 
-app.get("/threads/getList", threadsService2.threadsList)
-
 const start = async () => {
   try {
-    app.listen(process.env.SERVER_PORT, () => console.log("server is listening in PORT 8000"));
+    await myDataSource.initialize().then(() => console.log("Data Source has been initialized!"));
+
+    const port = process.env.SERVER_PORT;
+    app.listen(port, () => console.log(`server is listening in PORT ${port}`));
   } catch (error) {
     console.error(error);
   }
